Changing hook files with file name convention for consistency Refactor: following SR for the Edit submission files and moving them to the correct folder
