+++ apps/endatix-hub/app/(main)/forms/[formId]/submissions/[submissionId]/edit/page.tsx
@@ -4,7 +4,7 @@ import { Skeleton } from '@/components/ui/skeleton';
 import { getSubmissionDetailsUseCase } from '@/features/submissions/use-cases/get-submission-details.use-case';
 import { BackToSubmissionsButton } from '@/features/submissions/ui/details/back-to-submissions-button';
 import { Result } from '@/lib/result';
-import EditSubmission from './edit-submission';
+import EditSubmission from '@/features/submissions/ui/edit/edit-submission';
 
 type Params = {
   params: Promise<{
@@ -40,9 +40,7 @@ export default async function EditSubmissionPage({ params }: Params) {
 
   return (
     <Suspense fallback={<SubmissionDataSkeleton />}>
-      <EditSubmission 
-        submission={submission} 
-      />
+      <EditSubmission submission={submission} />
     </Suspense>
   );
 }
