+++ apps/endatix-hub/features/public-form/application/submission-queue/index.ts
@@ -1,1 +1,1 @@
-export { useSubmissionQueue } from './use-submission-queue';
+export { useSubmissionQueue } from './use-submission-queue.hook';
